+++ internal/pkg/message/policy.go
@@ -71,7 +71,7 @@ func (p *DefaultQueuePolicy) HandleNewHead(ctx context.Context, target PolicyTar
 				return err
 			}
 			for _, minedMsg := range secpMsgs {
-				removed, found, err := target.RemoveNext(ctx, minedMsg.Message.From, uint64(minedMsg.Message.CallSeqNum))
+				removed, found, err := target.RemoveNext(ctx, minedMsg.Message.From, minedMsg.Message.CallSeqNum)
 				if err != nil {
 					return err
 				}
